Use token centers to find positioning
Changes Marker.getImageDimensions to use the actual token object pulled
from the canvas. This is to conform the change of the type of the token
var in the move() and create() methods.

Changes Marker.getImageLocation to use the actual token and the
dimensions of the marker image. Using the actual token rather than
token.data allows access to Token.center allowing more straightforward
positioning of the marker.

Moves the move method to the base class since both versions have
identical logic.

Adds a getter for ratio to avoud having to pass it around. On the base
Marker class, returns fixed 1.5. StartMarker inherits, and TurnMarker
overrides to return the value from settings.

Fixes #6